type-bridge: checkpoint: integration-test buildout
 - added mkDatum() and (new)readDatum() interface to stellar contract class
 - renamed single-letter variant names to contain expressions of the
   (test-related) purposes for the fields
 - new test with most cases expressed in terms of expectations
   and most with drafts of api-level expectations. More refinement coming.
 - diagram for visualizing the structure of the type-bridge
   and how it fits together with dApp code and on-chain code
 - initial prototype for implementing runtime data conversions
   to match the expected API

TODO: adjust implementation approach to support documentation and
explicit materialization of type-bridging code.  Proxies are a cool
javascript feature but create indirection and unnecessary magic for
this use-case, given the existing rollup context in which we can
directly generate code exactly matching every dApp-specific type.  This
code will then be explicit, portable, easy to consume, audit, trace
with eyes or debugger, and with (automated) changes clearly traceable
to changes in the underlying data structures.