type-gen: Helios project: re-initialize when non-default Capo is found late
 - generates a new project object when the new Capo arrives (if there
   are other hlbundles registered & linked to the default capo bundle.

 - takes nearly 1 second each build, but it's proving difficult, as yet,
   to initialize the custom capo bundle first. TODO.

 - fixes guidance for defining a custom capo bundle

TODO: possibly defer type-generation until a custom capo is registered,
or until it's discovered there is none.  However, this would likely also
defer reporting of syntax problems in Helios scripts, so it may not be
any easy trade.  On the other hand, if dependent .hlbundles are relying on
dependency injection from the custom capo's defined modules, it may
actually be essential to take this approach.