checkpoint: enable on-chain deployment configurations
When building a dApp for deployment, its settings combine details from
the original Capo deployment, its current on-chain code and the deployment
environment.  By capturing these details, we can bundle ready-to-use
CBOR scripts for all our dApps' on-chain contracts.

This will be a per-environment capability; this version focuses on one
environment first.

 - added types for deployment details
 - changes signature for script-bundle creation (not including the CapoBundle)
   to optionally include the setup/deployment details needed for compiling
   and/or instantiating compiled programs from those details
 - modified code paths for instantiating Stellar-contract classes and
   their bundles: connectDelegate(), mustGetDelegate, addStrellaWithConfig,
   compileWithScriptParams(), bundle.createWith(); added bundle.withSetupDetails()
 - added diagram showing these code paths with types and notes
   - reflects current types and logic
 - light renaming of other types