checkpoint: multi-submitter working prototype
submits a chain of dependent txns reliably to the network, with retries,
progress monitoring, and (wip) multiple points of submission/resubmission

See included diagram

 - NEEDS minor upstream interface adjustments in Helios
 - tcx can be a facade/envelope for addlTxns & use batcher
 - batcher in setup arranges this-batch/last-batch semantics
 - TxSubmitMgr handles multiple txns and multiple underlying submitters
   - blockfrost submitter is best available single submitter for now

TODO: ogmios submitter is WIP; needs a hasTx or similar before it can
be as reliable as blockfrost.  The information can be found through
API calls, but because the chain is always shifting,
it's non-trivial. Needs a stateful list of recent network blocks
and txns in those blocks, kept fresh even in case of rollback