ONCHAIN: Change mint-delegate approval policy to allow Spent charter
... not only charter as ref-script.

The charter is needed at this moment in order to determine the mint-delegate
authority UUT that is required to get involved in the scenario.
The scenario allows a new UUT to be minted for a new-delegate-creation
moment when the charter data is ALSO being updated to reflect the presence
of that new delegate.  Required now and not earlier because Plomin hard-fork
disallows a RefInput from co-existing with the same UTXO as Input.

Auditors: this Spent Charter should always be spent based on its own
policy considerations, without the need for this code path (in the Minter)
to do any special checks.